Update look and feel of settings
