+++ Sources/App/Resources/en.lproj/Localizable.strings
@@ -551,6 +551,7 @@ Your actions and local configuration will still be available after logging in.";
 "settings.connection_section.cloudhook_url.title" = "Cloudhook URL";
 "settings.connection_section.connect_row.title" = "Connect";
 "settings.connection_section.connecting_via" = "Connected via";
+"settings.connection_section.add_server" = "Add Server";
 "settings.connection_section.details" = "Details";
 "settings.connection_section.error_enabling_notifications.message" = "There was an error enabling notifications. Please try again.";
 "settings.connection_section.error_enabling_notifications.title" = "Error enabling notifications";
@@ -635,6 +636,7 @@ Your actions and local configuration will still be available after logging in.";
 "settings.discovery_section.requiresPassword" = "Requires password";
 "settings.event_log.title" = "Event Log";
 "settings.general_settings_button.title" = "General";
+"settings.debugging.title" = "Debugging";
 "settings.location_history.description" = "No Location History";
 "settings.location_history.detail.explanation" = "The purple circle is your location and its accuracy. Blue circles are your zones. You are inside a zone if the purple circle overlaps a blue circle. Orange circles are additional regions used for sub-100 m zones.";
 "settings.location_history.empty" = "No Location History";
@@ -953,3 +955,4 @@ Thank you.\
 "widgets.actions.not_configured" = "No Actions Configured";
 "widgets.actions.title" = "Actions";
 "yes_label" = "Yes";
+"help_label" = "Help";
