+++ Sources/Shared/Settings/SettingsStore.swift
@@ -105,41 +105,6 @@ public class SettingsStore {
         }
     }
 
-    private var testAuthenticatedUser: AuthenticatedUser?
-
-    public var authenticatedUser: AuthenticatedUser? {
-        get {
-            if Current.isRunningTests {
-                return testAuthenticatedUser
-            }
-
-            guard let userData = ((try? keychain.getData("authenticatedUser")) as Data??),
-                  let unwrappedData = userData else {
-                return nil
-            }
-
-            return try? JSONDecoder().decode(AuthenticatedUser.self, from: unwrappedData)
-        }
-        set {
-            if Current.isRunningTests {
-                testAuthenticatedUser = newValue
-                return
-            }
-
-            guard let info = newValue else {
-                keychain["authenticatedUser"] = nil
-                return
-            }
-
-            do {
-                let data = try JSONEncoder().encode(info)
-                try keychain.set(data, key: "authenticatedUser")
-            } catch {
-                assertionFailure("Error while saving authenticated user info: \(error)")
-            }
-        }
-    }
-
     public var pushID: String? {
         get {
             prefs.string(forKey: "pushID")
